requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
claude-agent-sdk>=0.1
python-telegram-bot[job-queue]>=21.0,<22
python-dotenv>=1.0,<2
aiosqlite>=0.20,<1

# ── Dashboard (Phase 3) ────────────────────────────────────────────────────────
fastapi>=0.115,<1
uvicorn[standard]>=0.32,<1
httpx>=0.27,<1

# ── Memory (Phase 4) ──────────────────────────────────────────────────────────
rank-bm25>=0.2,<1
watchdog>=4.0,<5
# Optional vector embeddings (set MEMORY_EMBEDDING_PROVIDER=local to enable):
# sentence-transformers>=3.0,<4
# sqlite-vec

# ── Media, voice & documents (Phase 8/9) ──────────────────────────────────────
# Voice transcription via OpenAI Whisper API (requires OPENAI_API_KEY):
openai>=1.0,<2
# OR local Whisper (CPU, no API key, ~500 MB model download on first run):
# faster-whisper>=1.0,<2
# PDF text extraction (optional — install to enable PDF uploads):
# pypdf>=4.0,<5

# ── Discord channel (Phase 8) ─────────────────────────────────────────────────
discord.py>=2.3,<3
